requirements.txt
greenlet==3.0.1
numpy==1.26.2
pandas==2.1.3
playwright==1.40.0
playwright-stealth==1.0.6
psycopg2-binary==2.9.9
pyee==11.0.1
python-dateutil==2.8.2
pytz==2023.3.post1
selectolax==0.3.17
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
tzdata==2023.3
